+++ docs/changelog.md
@@ -12,6 +12,7 @@ category: Administration
 ### [Unreleased]
 
 * fix: always open documentation in subtasks dialog in new window
+* feat: ensure compatibility with Jira 9.15.0, 9.12.5, 9.4.18
 
 ### [24.01.0] - 2024-01-27