+++ docs/changelog.md
@@ -15,7 +15,8 @@ category: Administration
 * feat: ensure compatibility with Jira 9.15.0, 9.12.5, 9.4.18
 * feat: documentation link in dialog footer is always visible (CASUP-110)
 * feat: ensure compatibility with Jira 9.16.0, 9.12.9, 9.4.22
-* fix: robustness of template import for non-UTF8 characters 
+* fix: robustness of template import for non-UTF8 characters
+* feat: ensure compatibility with Jira 9.17.0
 
 ### [24.01.0] - 2024-01-27